Moved code to class
