+++ tests/pose_slam/test_pose_slam.py
@@ -51,75 +51,12 @@ def read_fastlio_result(poses_txt_path: str) -> Tuple[List[Pose3], np.ndarray]:
     return [pose_of_vector(pose) for pose in poses], timestamps
 
 
-def angle(R1, R2):
-    """Calculate angle between two rotations, in degrees."""
-    return np.degrees(np.linalg.norm(R1.logmap(R2)))
-
-
-def difference(P1, P2):
-    """Calculate the translation and angle differences of two poses.
-    P1, P2: Pose3
-    Return:
-        distance: translation difference
-        angle: angular difference
-    """
-    # TODO(frank): clean this up
-    t1 = P1.translation()
-    t2 = P2.translation()
-    R1 = P1.rotation()
-    R2 = P2.rotation()
-    R1_2 = R1.compose(R2.inverse())
-    t1_ = R1_2.rotate(t2)
-    # t1_2 = t1 - R1_2*t2
-    distance = np.linalg.norm(t1 - t1_)
-    angle_ = angle(R1, R2)
-    return distance, angle_
-
-
 def create_initial_estimate(poses: List[Pose3]):
     """Create initial estimate"""
     initial_estimate = Values()
     for i, pose_vector in enumerate(poses):
         initial_estimate.insert(i, pose_vector)
     return initial_estimate
-
-
-def filtered_pose_priors(
-    constraints: List[Constraint], poses: List[Pose3], add_backbone=True
-) -> Dict[Tuple[int, int], PosePrior]:
-    """Generate relative pose priors from constraints and initial_estimate by filtering heavily. Optionally add back bone."""
-
-    relative_pose_priors: Dict[Tuple[int, int], PosePrior] = {}
-
-    # Create pose constraints using Bayesian ICP
-    for constraint in constraints:
-        a, b = constraint.a, constraint.b
-        aTb, cov = constraint.aTb, constraint.cov
-        predicted_aTb = poses[a].between(poses[b])
-        trans_diff, rot_diff = difference(aTb, predicted_aTb)
-        inlier = (trans_diff <= 0.04) and (rot_diff <= 5)
-        if not inlier:
-            continue
-        try:
-            info = np.linalg.inv(cov)
-            if np.isnan(info).any():
-                continue
-            else:
-                relative_pose_priors[(a, b)] = PosePrior(aTb, cov, PosePriorType.SOFT_CONSTRAINT)
-        except np.linalg.LinAlgError:
-            continue
-
-    # Create loose odometry factors for backbone.
-    if add_backbone:
-        backbone_cov = np.diag(np.array([1, 1, 1, np.deg2rad(30.0), np.deg2rad(30.0), np.deg2rad(30.0)]))
-        for i in range(len(poses) - 1):
-            a, b = i, i + 1
-            if (a, b) not in relative_pose_priors:
-                aTb = poses[a].between(poses[b])
-                relative_pose_priors[(a, b)] = PosePrior(aTb, backbone_cov, PosePriorType.SOFT_CONSTRAINT)
-
-    # Output the resulting poses and pose constraints
-    return relative_pose_priors
 
 
 def generate_pose_graph(poses: List[Pose3], relative_pose_priors: Dict[Tuple[int, int], PosePrior]):
@@ -187,8 +124,7 @@ class TestPoseSlam(GtsamTestCase):
         self.assertEqual(len(constraints), 10328)
 
         initial_estimate = create_initial_estimate(poses)
-        # graph, initial_estimate = self.slam.generate_pose_graph(constraints, poses)
-        relative_pose_priors = filtered_pose_priors(constraints, poses)
+        relative_pose_priors = self.slam.filtered_pose_priors(constraints, poses)
         graph = generate_pose_graph(poses, relative_pose_priors)
         self.assertEqual(graph.size(), 10062)
         self.assertAlmostEqual(graph.error(initial_estimate), 35111.87458699957)
